Documentation
Help get started using this application with our documentation, which helps with getting started and going through the basics of the installer. The main application itself is self-explanatory and should be easy to navigate. Any questions, concerns, or want to report a bug? Click on the link at the bottom of the documentation to get your request fulfilled. 
Thank you for viewing our application. — Alen Jo (Owner of GeekDictionary Software)